Refactoring to blueprints
