+++ script_runner/blueprints/base_app_bp.py
@@ -1,0 +1,13 @@
+from flask import (
+    Blueprint,
+    Response,
+    jsonify,
+    make_response,
+)
+
+base_app_bp = Blueprint("base", __name__)
+
+
+@base_app_bp.route("/health")
+def health() -> Response:
+    return make_response(jsonify({"status": "ok"}))
